Moved the 'URL' attribute of Entries to a property
Storing the URLs in the database causes errors on save for pages with
long titles (close to 255 characters) because the generated URLs would
end up being longer than 255 characters when combined with path
information. Moving the URL attribute to a property bypassess this issue
by not attempting to put it in the database. We could also address this
by increasing the max_length of the URL field but there doesn't seem to
be any reason for storing this information on the Entry other than for
convenience, so the property should suffice.